Project 1 ENDG 310
In this project, you will create a visual graphic of data from a dataset of your choosing. The dataset may be of any topic, however, it must meet the following criteria:
Come from a trustworthy and verifiable public source.
Be properly cited in your project.
Contain at least two columns of data and consist of at least 500 rows.
With this data, you need to create a visual that summarizes the data and presents some insight.Simply plotting the data is not sufficient - you must manipulate the data to derive additionalinformation (a running average, a trendline, annotations, etc). The visual must be easy tounderstand.